Added test.py attempting to understand how this all works, and removed inputdata entries.
test.py is essentially a prettier clone of main.py for my own benefit,
but I can see clearly now how we could define a class with methods to
be a nice facade to this. I removed the input data because I figured we
didn't need it here. Maybe we sholud .gitignore it?